render fetched data
