+++ pages/_app.js
@@ -1,4 +1,6 @@
 import GlobalStyle from "../styles";
+import useSWR from "swr";
+import ArtworksPreview from "@/components/ArtworksPreview";
 
 export default function App({ Component, pageProps }) {
   return (
